perf(stats): lazy-load engagement counter island
