+++ src/lib/domains/PageInteraction/StatsCounter.tsx
@@ -1,29 +1,20 @@
+/**
+ * Lightweight stats island shell.
+ * Page-view recording stays on the initial hydration path so cached pages still
+ * count visits quickly, while the below-the-fold stats UI and its React Query /
+ * icon / presence dependencies load as a lazy island.
+ */
 "use client";
 
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useLocation } from "@tanstack/react-router";
-import { useServerFn } from "@tanstack/react-start";
-import { type ReactNode } from "react";
-import { FaEye, FaHeart, FaRegCircleUser, FaRegHeart } from "react-icons/fa6";
+import { lazy, Suspense } from "react";
 
-import { IS_CI, IS_DEV } from "@/config";
-import { useLiveViewerCount } from "@/lib/components/useLiveViewerCount";
-import { cn, normalizePathname } from "@/lib/helpers/utils";
+import { IS_CI } from "@/config";
+import { normalizePathname } from "@/lib/helpers/utils";
 
-import { decrementLikeServerFn, incrementLikeServerFn, type LikeCount } from "./LikeButton.server";
-import { type PageMetrics } from "./Metrics.server";
-import {
-	getLikeButtonClassName,
-	getLikeHeartIconClassName,
-	getLikeLoadingHeartIconClassName,
-	getLikeMutationAction,
-	type LikeMutationAction,
-} from "./StatsCounter.helpers";
-import { pageMetricsQueryKey, usePageMetrics } from "./usePageMetrics";
 import { useRecordPageView } from "./useRecordPageView";
-import { fetchViewCountServerFn, type PageViewCount } from "./ViewsCounter.server";
 
-type StatsCounterProps = {
+export type StatsCounterProps = {
 	slug?: string;
 	page_type?: "post" | "page";
 	hidden?: boolean;
@@ -31,503 +22,20 @@ type StatsCounterProps = {
 	variant?: "views" | "engagement";
 };
 
-const statsListClassName = "m-0 flex min-h-5 flex-wrap items-center justify-center gap-4 text-sm";
-const statItemClassName = "flex items-center justify-center gap-1.5";
-const statValueClassName = "inline-block min-w-[2ch] text-left tabular-nums";
-const statTooltipTriggerClassName = "group relative";
-const statTooltipBubbleClassName =
-	"pointer-events-none absolute bottom-full left-1/2 z-50 mb-2 w-max max-w-64 -translate-x-1/2 rounded-global border border-solid border-foreground/15 bg-background px-2.5 py-1.5 text-center text-xs leading-snug text-foreground opacity-0 shadow-lg transition-opacity duration-100 group-hover:opacity-100 group-focus-within:opacity-100";
+const StatsCounterView = lazy(() =>
+	import("./StatsCounter.view").then((module) => ({ default: module.StatsCounterView })),
+);
 
-export const StatsCounter = ({
-	slug,
-	page_type = "page",
-	hidden = false,
-	disabled = IS_CI,
-	variant = "engagement",
-}: StatsCounterProps) => {
+export const StatsCounter = (props: StatsCounterProps) => {
 	const { pathname } = useLocation();
-	const normalized_slug = slug ?? pathname;
-	const normalizedPathname = normalizePathname(normalized_slug);
+	const disabled = props.disabled ?? IS_CI;
+	const normalizedPathname = normalizePathname(props.slug ?? pathname);
+
 	useRecordPageView(normalizedPathname, disabled);
 
-	if (variant === "engagement") {
-		return (
-			<div
-				className={cn(
-					"mx-auto mt-auto mb-5 w-full flex-row flex-wrap items-center justify-center gap-2 pt-40",
-					hidden ? "hidden" : "flex",
-				)}
-			>
-				<StatsSentence
-					normalizedPathname={normalizedPathname}
-					disabled={disabled}
-					page_type={page_type}
-				/>
-			</div>
-		);
-	}
-
 	return (
-		<div
-			className={cn(
-				"mx-auto mt-auto mb-5 w-full flex-row flex-wrap items-center justify-center gap-2 pt-40",
-				hidden ? "hidden" : "flex",
-			)}
-		>
-			<StandaloneViews
-				normalizedPathname={normalizedPathname}
-				disabled={disabled}
-				page_type={page_type}
-			/>
-		</div>
+		<Suspense fallback={null}>
+			<StatsCounterView {...props} normalizedPathname={normalizedPathname} disabled={disabled} />
+		</Suspense>
 	);
 };
-
-type StatsSentenceProps = {
-	normalizedPathname: string;
-	disabled: boolean;
-	page_type: StatsCounterProps["page_type"];
-};
-
-const StatsSentence = ({ normalizedPathname, disabled, page_type }: StatsSentenceProps) => {
-	const { data, isError, isLoading } = usePageMetrics(normalizedPathname, disabled);
-	const noun = page_type === "post" ? "post" : "page";
-
-	if (isLoading) {
-		return (
-			<StatsList noun={noun} isBusy statusLabel={`Getting stats for this ${noun}`}>
-				<MetricSkeleton
-					icon={<FaEye aria-hidden="true" focusable={false} className="size-5 text-primary" />}
-					label="Views"
-					valueWidthClassName="w-[5ch]"
-				/>
-				<MetricSkeleton icon={<LoadingLikeHeart />} label="Likes" valueWidthClassName="w-[2ch]" />
-				<LiveStat />
-			</StatsList>
-		);
-	}
-
-	if (isError || data?.view_count === undefined) {
-		return (
-			<p className="m-0 min-h-5 text-sm" role="status">
-				Stats unavailable right now
-			</p>
-		);
-	}
-
-	return (
-		<StatsList noun={noun}>
-			<ViewsStat value={data.view_count} noun={noun} />
-			<LikeStat
-				normalizedPathname={normalizedPathname}
-				disabled={disabled}
-				metrics={data}
-				noun={noun}
-			/>
-			<LiveStat />
-		</StatsList>
-	);
-};
-
-const StatsList = ({
-	noun,
-	isBusy = false,
-	statusLabel,
-	children,
-}: {
-	noun: "post" | "page";
-	isBusy?: boolean;
-	statusLabel?: string;
-	children: ReactNode;
-}) => (
-	<>
-		{statusLabel ? <output className="sr-only">{statusLabel}</output> : null}
-		<dl
-			className={statsListClassName}
-			aria-label={statusLabel ?? `${noun} engagement stats`}
-			aria-busy={isBusy}
-		>
-			{children}
-		</dl>
-	</>
-);
-
-const MetricSkeleton = ({
-	icon,
-	label,
-	valueWidthClassName,
-}: {
-	icon: ReactNode;
-	label: string;
-	valueWidthClassName: string;
-}) => (
-	<div className={statItemClassName}>
-		<dt className="sr-only">{label}</dt>
-		<dd className="m-0 inline-flex min-h-5 items-center gap-1.5">
-			{icon}
-			<span
-				aria-hidden="true"
-				className={cn("h-4 animate-pulse rounded-full bg-foreground/15", valueWidthClassName)}
-			/>
-			<span className="sr-only">Loading {label.toLowerCase()}</span>
-		</dd>
-	</div>
-);
-
-const StatTooltipBubble = ({ children }: { children: ReactNode }) => (
-	<span aria-hidden="true" className={statTooltipBubbleClassName}>
-		{children}
-		<span
-			aria-hidden="true"
-			className="absolute top-full left-1/2 size-2 -translate-x-1/2 -translate-y-1/2 rotate-45 border-r border-b border-solid border-foreground/15 bg-background"
-		/>
-	</span>
-);
-
-const LoadingLikeHeart = () => (
-	<span className={getLikeLoadingHeartIconClassName()}>
-		<FaHeart aria-hidden="true" focusable={false} className="size-5 text-primary/30" />
-		<FaHeart
-			aria-hidden="true"
-			focusable={false}
-			className="pointer-events-none absolute inset-0 size-5 animate-like-heart-shimmer text-secondary motion-reduce:animate-pulse"
-		/>
-	</span>
-);
-
-const ViewsStat = ({ value, noun }: { value: number; noun: "post" | "page" }) => {
-	const formattedValue = value.toLocaleString();
-	const label = `${formattedValue} views`;
-	const tooltipContent = `This ${noun} has ${label}.`;
-
-	return (
-		<div className={statItemClassName}>
-			<dt className="sr-only">Views</dt>
-			<dd
-				className={cn("m-0 inline-flex items-center gap-1.5", statTooltipTriggerClassName)}
-				aria-label={label}
-			>
-				<FaEye aria-hidden="true" focusable={false} className="size-5 text-primary" />
-				<span aria-hidden="true" className={statValueClassName}>
-					{formattedValue}
-				</span>
-				<StatTooltipBubble>{tooltipContent}</StatTooltipBubble>
-			</dd>
-		</div>
-	);
-};
-
-const LiveStat = () => {
-	const { count, connected } = useLiveViewerCount();
-
-	if (count === null) {
-		return (
-			<MetricSkeleton
-				icon={
-					<FaRegCircleUser
-						aria-hidden="true"
-						focusable={false}
-						className="size-5 rounded-full text-primary"
-					/>
-				}
-				label="Live viewers across the site"
-				valueWidthClassName="w-[2ch]"
-			/>
-		);
-	}
-
-	const formattedCount = count.toLocaleString();
-	const liveViewerUnit = count === 1 ? "live viewer" : "live viewers";
-	const label = `${formattedCount} ${liveViewerUnit} across the site`;
-	const tooltipContent = `There ${count === 1 ? "is" : "are"} ${label} right now.`;
-
-	return (
-		<div className={statItemClassName}>
-			<dt className="sr-only">Live viewers across the site</dt>
-			<dd
-				className={cn("m-0 inline-flex items-center gap-1.5", statTooltipTriggerClassName)}
-				aria-label={label}
-			>
-				<span
-					aria-hidden="true"
-					className="relative inline-flex size-5 items-center justify-center"
-				>
-					{connected ? (
-						<span className="absolute inline-flex h-full w-full animate-ping rounded-full bg-primary opacity-75 animate-duration-[1000ms] motion-reduce:animate-none" />
-					) : null}
-					<FaRegCircleUser
-						aria-hidden="true"
-						focusable={false}
-						className="relative inline-flex size-5 rounded-full text-primary"
-					/>
-				</span>
-				<span aria-hidden="true" className={statValueClassName}>
-					{formattedCount}
-				</span>
-				<StatTooltipBubble>{tooltipContent}</StatTooltipBubble>
-			</dd>
-		</div>
-	);
-};
-
-const LikeStat = ({
-	normalizedPathname,
-	disabled,
-	metrics,
-	noun,
-}: {
-	normalizedPathname: string;
-	disabled: boolean;
-	metrics: PageMetrics;
-	noun: "post" | "page";
-}) => {
-	const queryClient = useQueryClient();
-	const queryKey = pageMetricsQueryKey(normalizedPathname);
-
-	const incrementLikeCount = useServerFn<() => Promise<LikeCount>>(() =>
-		incrementLikeServerFn({ data: { slug: normalizedPathname, disabled } }),
-	);
-	const decrementLikeCount = useServerFn<() => Promise<LikeCount>>(() =>
-		decrementLikeServerFn({ data: { slug: normalizedPathname, disabled } }),
-	);
-	const {
-		mutate,
-		isPending,
-		variables: pendingAction,
-	} = useMutation({
-		mutationFn: (action: LikeMutationAction) =>
-			action === "unlike" ? decrementLikeCount() : incrementLikeCount(),
-		onMutate: async (action) => {
-			await queryClient.cancelQueries({ queryKey });
-
-			const previous = queryClient.getQueryData<PageMetrics>(queryKey);
-			if (!previous) {
-				return { previous };
-			}
-
-			if (action === "like" && previous.hasLiked) {
-				return { previous };
-			}
-			if (action === "unlike" && !previous.hasLiked) {
-				return { previous };
-			}
-
-			queryClient.setQueryData<PageMetrics>(queryKey, (old) =>
-				old
-					? {
-							...old,
-							likes: action === "unlike" ? Math.max(0, old.likes - 1) : old.likes + 1,
-							hasLiked: action === "like",
-						}
-					: old,
-			);
-
-			return { previous };
-		},
-		onError: (_error, _variables, context) => {
-			if (context?.previous) {
-				queryClient.setQueryData<PageMetrics>(queryKey, context.previous);
-			}
-		},
-		onSuccess: (likeCount) => {
-			queryClient.setQueryData<PageMetrics>(queryKey, (old) =>
-				old ? { ...old, ...likeCount } : old,
-			);
-		},
-	});
-
-	const like_count = metrics.likes;
-	const hasLiked = metrics.hasLiked;
-	const readOnly = metrics.readOnly ?? false;
-	const canUnlike = hasLiked && IS_DEV;
-	const nextLikeAction = getLikeMutationAction({ hasLiked, isDev: IS_DEV });
-	const is_disabled = disabled || isPending || readOnly || (hasLiked && !canUnlike);
-	const formatted_like_count = like_count.toLocaleString();
-	const likeUnit = like_count === 1 ? "like" : "likes";
-	const likeSummary = `${formatted_like_count} ${likeUnit}`;
-	const likeActionLabel = hasLiked
-		? canUnlike
-			? `Unlike this ${noun}`
-			: `You liked this ${noun}`
-		: `Like this ${noun}`;
-	const likeTooltipText = readOnly
-		? `This ${noun} has ${likeSummary}. Likes are read-only here.`
-		: hasLiked
-			? canUnlike
-				? `This ${noun} has ${likeSummary}. Click the heart to unlike it in local dev.`
-				: `This ${noun} has ${likeSummary}. You have liked it.`
-			: `This ${noun} has ${likeSummary}. Click the heart to like it.`;
-	const pendingVerb = pendingAction === "unlike" ? "unlike" : "like";
-	const likeButtonLabel = isPending
-		? `Saving ${pendingVerb} for this ${noun}. ${likeSummary}`
-		: `${likeActionLabel}. ${likeSummary}`;
-	const likeButtonTitle = isPending
-		? `Saving ${pendingVerb} for this ${noun} — ${likeSummary}`
-		: `${likeActionLabel} — ${likeSummary}`;
-
-	return (
-		<div className={statItemClassName}>
-			<dt className="sr-only">Likes</dt>
-			<dd className={cn("m-0 inline-flex items-center gap-1.5", statTooltipTriggerClassName)}>
-				<button
-					type="button"
-					onClick={() => mutate(nextLikeAction)}
-					aria-label={likeButtonLabel}
-					aria-pressed={hasLiked}
-					className={getLikeButtonClassName({ hasLiked })}
-					disabled={is_disabled}
-				>
-					{hasLiked ? (
-						<span className={getLikeHeartIconClassName({ isPending })}>
-							<FaHeart
-								aria-hidden="true"
-								focusable={false}
-								className={cn("size-5", isPending && "text-primary/45")}
-							/>
-							{isPending ? (
-								<FaHeart
-									aria-hidden="true"
-									focusable={false}
-									className="pointer-events-none absolute inset-0 size-5 animate-like-heart-shimmer text-secondary motion-reduce:animate-pulse"
-								/>
-							) : null}
-						</span>
-					) : (
-						<FaRegHeart aria-hidden="true" focusable={false} className="size-5" />
-					)}
-				</button>
-				<span aria-hidden="true" className={statValueClassName}>
-					{formatted_like_count}
-				</span>
-				<StatTooltipBubble>{isPending ? likeButtonTitle : likeTooltipText}</StatTooltipBubble>
-			</dd>
-		</div>
-	);
-};
-
-const StandaloneViews = ({ normalizedPathname, disabled, page_type }: InnerViewsProps) => {
-	const fetchViewCount = useServerFn<() => Promise<PageViewCount>>(() =>
-		fetchViewCountServerFn({ data: { slug: normalizedPathname, disabled } }),
-	);
-	const { data, isError, isLoading } = useQuery({
-		queryFn: fetchViewCount,
-		queryKey: [normalizedPathname, "get-views"],
-		staleTime: 1000 * 30,
-		// This RPC is now read-only (the increment moved to the Worker
-		// document-request path), but we still avoid background replays so a stale
-		// window doesn't churn the counter reads.
-		refetchOnWindowFocus: false,
-		refetchOnReconnect: false,
-		retry: false,
-	});
-	return (
-		<ViewsCopy
-			isError={isError}
-			isLoading={isLoading}
-			view_count={data?.view_count}
-			page_type={page_type}
-		/>
-	);
-};
-
-type InnerViewsProps = {
-	normalizedPathname: string;
-	disabled: boolean;
-	page_type: StatsCounterProps["page_type"];
-};
-
-const ViewsCopy = ({
-	isError,
-	isLoading,
-	view_count,
-	page_type,
-}: {
-	isError: boolean;
-	isLoading: boolean;
-	view_count: number | undefined;
-	page_type: StatsCounterProps["page_type"];
-}) => {
-	if (isLoading) {
-		return (
-			<p
-				className="m-0 inline-flex min-h-7 items-center gap-2 text-sm"
-				aria-label={`Getting view count for this ${page_type}`}
-				role="status"
-			>
-				<span aria-hidden="true" className="h-5 w-48 animate-pulse rounded-full bg-foreground/15" />
-			</p>
-		);
-	}
-
-	if (isError || view_count === undefined) {
-		return <p className="m-0 text-sm">View count unavailable right now</p>;
-	}
-
-	return <p className="m-0 text-sm">{getViewCountCopy(view_count, page_type)}</p>;
-};
-
-function getViewCountCopy(view_count: number, page_type: StatsCounterProps["page_type"]) {
-	switch (view_count) {
-		case 0:
-			return "No views yet. Wait what, HOW? 🤔";
-		case 1:
-			return "This page has been viewed only once. That's a lot of views!";
-		case 69:
-			return (
-				<>
-					This {page_type} has been viewed <MetricValue>{view_count.toLocaleString()}</MetricValue>{" "}
-					times. Nice.
-				</>
-			);
-		case 420:
-			return (
-				<>
-					This {page_type} has been viewed <MetricValue>{view_count.toLocaleString()}</MetricValue>{" "}
-					times. Hehe.
-				</>
-			);
-
-		default: {
-			if (view_count > 10000) {
-				return (
-					<>
-						This {page_type} has been viewed{" "}
-						<MetricValue>{view_count.toLocaleString()}</MetricValue> times. Holy crap.
-					</>
-				);
-			}
-			if (view_count > 1000) {
-				return (
-					<>
-						This {page_type} has been viewed{" "}
-						<MetricValue>{view_count.toLocaleString()}</MetricValue> times. Holy crap. 🤯
-					</>
-				);
-			}
-			if (view_count > 100) {
-				return (
-					<>
-						This {page_type} has been viewed{" "}
-						<MetricValue>{view_count.toLocaleString()}</MetricValue> times. Wow.
-					</>
-				);
-			}
-			return (
-				<>
-					This {page_type} has been viewed <MetricValue>{view_count.toLocaleString()}</MetricValue>{" "}
-					times
-				</>
-			);
-		}
-	}
-}
-const MetricValue = ({ children, className }: { children: string; className?: string }) => (
-	<span
-		className={cn(
-			"rounded-global border-2 border-solid border-primary bg-background p-1 font-mono text-base text-primary tabular-nums transition-colors",
-			className,
-		)}
-	>
-		{children}
-	</span>
-);
