+++ src/lib/domains/PageInteraction/usePageMetrics.ts
@@ -4,6 +4,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useServerFn } from "@tanstack/react-start";
 
 import { fetchPageMetricsServerFn, type PageMetrics } from "./Metrics.server";
+import { waitForPageViewRecord } from "./useRecordPageView";
 
 export function pageMetricsQueryKey(normalizedPathname: string) {
 	return [normalizedPathname, "metrics"] as const;
@@ -17,7 +18,10 @@ export function usePageMetrics(normalizedPathname: string, disabled: boolean) {
 	);
 
 	return useQuery({
-		queryFn: fetchMetrics,
+		queryFn: async () => {
+			await waitForPageViewRecord(normalizedPathname);
+			return fetchMetrics();
+		},
 		queryKey: pageMetricsQueryKey(normalizedPathname),
 		enabled: !disabled,
 		staleTime: 1000 * 30,
